refactor(api): move oidc-provider controllers
to oidc-provider directory
