+++ api/src/identity-access-management/application/routes.js
@@ -1,5 +1,5 @@
-import { oidcProviderAdminRoutes } from './oidc-provider.admin.route.js';
-import { oidcProviderRoutes } from './oidc-provider.route.js';
+import { oidcProviderAdminRoutes } from './oidc-provider/oidc-provider.admin.route.js';
+import { oidcProviderRoutes } from './oidc-provider/oidc-provider.route.js';
 import { samlRoutes } from './saml.route.js';
 import { tokenRoutes } from './token.route.js';
 import { userAdminRoutes } from './user.admin.route.js';
